--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -37,6 +37,7 @@ Imports:
     Formula (>= 1.2-4),
     ggplot2 (>= 3.2.0),
     lubridate (>= 1.7.8),
+    numDeriv (>= 2016.8-1.1),
     Matrix (>= 1.2-17),
     MASS,
     optimx (>= 2019-12.02),