Use `numDeriv::hessian` which is also used by `optimx` internally. It is therefore already an installed dependency.

**Benchmark**
- keeping track of calls to LL for correlation only using a global variable HESSIAN_COUNTER that is incremented in the correlation interlayer


CDNOW
```
### Setup
clv.data.cdnow <- clvdata(data.transactions = cdnow, date.format = "ymd", time.unit = "w", estimation.split = 37)
HESSIAN_COUNTER <- 0
microbenchmark::microbenchmark(pnbd(clv.data.cdnow, use.cor = T, optimx.args = list(itnmax=5)), times=10)

### Old hessian
Unit: seconds
                                                              expr      min       lq    mean   median       uq      max neval
 pnbd(clv.data.cdnow, use.cor = T, optimx.args = list(itnmax = 5)) 3.213403 3.231136 3.25158 3.248915 3.271715 3.296059    10
HESSIAN_COUNTER=1722   # Single optimization

### New hessian
Unit: milliseconds
                                                              expr      min       lq     mean   median       uq      max neval
 pnbd(clv.data.cdnow, use.cor = T, optimx.args = list(itnmax = 5)) 413.3427 417.9993 425.7277 422.4331 427.7635 458.0337    10
HESSIAN_COUNTER=122   # Single optimization
```

DYNCOV
```
### Setup
clv.data.dyn.cov <- clvdata(data.transactions = apparelTrans, date.format = "ymd", time.unit = "w", estimation.split = 40)
HESSIAN_COUNTER <- 0
microbenchmark::microbenchmark(pnbd(clv.data.dyn.cov, use.cor = T, optimx.args = list(itnmax=5)), times=1)

### Old hessian
Unit: seconds
                                                                expr      min       lq     mean   median       uq      max neval
 pnbd(clv.data.dyn.cov, use.cor = T, optimx.args = list(itnmax = 5)) 178.5845 178.5845 178.5845 178.5845 178.5845 178.5845     1
HESSIAN_COUNTER=8010   # Single optimization

### New hessian
Unit: seconds
                                                                expr      min       lq     mean   median       uq      max neval
 pnbd(clv.data.dyn.cov, use.cor = T, optimx.args = list(itnmax = 5)) 14.73023 14.73023 14.73023 14.73023 14.73023 14.73023     1
HESSIAN_COUNTER=533   # Single optimization
```

**Summary**
cdnow
- 3.25s -> 0.42s -87%
- 1777 -> 122 calls to hessian:  -93%

Dyncov
- 178s-> 14.7s: -92%
- 8010 -> 533 calls to hessian: -93%

The calls to hessian are obviously dependent on the number of parameters but the time savings are not (-93% in both cases)